feat: review english strings; add shared list

+++ app/src/main/java/com/alvarengadev/marketplacelist/ui/fragments/cart/CartFragment.kt
@@ -1,5 +1,6 @@
 package com.alvarengadev.marketplacelist.ui.fragments.cart
 
+import android.content.Intent
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -18,18 +19,19 @@ import com.alvarengadev.marketplacelist.ui.fragments.cart.adapter.OnClickItemLis
 import com.alvarengadev.marketplacelist.ui.fragments.cart.dialog.feature.FeatureClearCartDialog
 import com.alvarengadev.marketplacelist.utils.Constants
 import com.alvarengadev.marketplacelist.utils.PreferencesManager
+import com.alvarengadev.marketplacelist.utils.TextFormatter
 import com.alvarengadev.marketplacelist.utils.extensions.createSnack
-import com.alvarengadev.marketplacelist.utils.extensions.goneWithoutAnimation
-import com.alvarengadev.marketplacelist.utils.extensions.visibilityWithoutAnimation
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
 class CartFragment : Fragment(R.layout.fragment_cart), ObserverListEmpty {
 
-    private var preferencesManager: PreferencesManager? = null
     private var _binding: FragmentCartBinding? = null
     private val binding get() = _binding!!
     private val cartViewModel: CartViewModel by viewModels()
+
+    private var preferencesManager: PreferencesManager? = null
+    private var intentSharedCart: Intent? = null
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -43,9 +45,8 @@ class CartFragment : Fragment(R.layout.fragment_cart), ObserverListEmpty {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        preferencesManager = PreferencesManager.instance
-        setupComponents()
         setupEvents()
+        setupComponents()
         listenToRegistrationViewModelEvents()
     }
 
@@ -68,9 +69,19 @@ class CartFragment : Fragment(R.layout.fragment_cart), ObserverListEmpty {
         ibMore.setOnClickListener {
             val popupMenu = context?.let { context -> PopupMenu(context, it) }
             popupMenu?.menuInflater?.inflate(R.menu.menu_popup_cart, popupMenu.menu)
-            popupMenu?.setOnMenuItemClickListener{
-                cartViewModel.deleteAllItemsFromDatabase()
-                true
+            popupMenu?.setOnMenuItemClickListener {
+                when (it.itemId) {
+                    R.id.sharedListMenu -> {
+                        intentSharedCart?.let { intent ->
+                            intent.type = "text/html"
+                            startActivity(Intent.createChooser(intent, "Share to"))
+                        }
+                    }
+                    R.id.clearCartMenu -> {
+                        cartViewModel.deleteAllItemsFromDatabase()
+                    }
+                }
+                false
             }
             popupMenu?.show()
         }
@@ -90,11 +101,11 @@ class CartFragment : Fragment(R.layout.fragment_cart), ObserverListEmpty {
         with(cartViewModel) {
             registrationStateEvent.observe(viewLifecycleOwner) { registrationState ->
                 if (registrationState is CartViewModel.CartListState.LoadingList) {
-                    pbLoadingList.visibilityWithoutAnimation()
+                    binding.cartViewFlipper.displayedChild = VIEW_FLIPPER_LOADING
                 }
 
                 if (registrationState is CartViewModel.CartListState.ListEmpty) {
-                    showList(false)
+                    binding.cartViewFlipper.displayedChild = VIEW_FLIPPER_EMPTY
                     footerCart.setCartValue()
                 }
 
@@ -115,12 +126,14 @@ class CartFragment : Fragment(R.layout.fragment_cart), ObserverListEmpty {
                         })
                     }
 
-
                     rcyCartItem.adapter = adapterListCart
-
-                    showList(true)
+                    binding.cartViewFlipper.displayedChild = VIEW_FLIPPER_LIST
                     viewDialogAlertFeatureClearCart()
                     footerCart.setCartValue(registrationState.total)
+                    intentSharedCart?.putExtra(
+                        Intent.EXTRA_TEXT,
+                        context?.let { TextFormatter.messageSharedList(it, registrationState.listItems) }
+                    )
                 }
 
                 if (registrationState is CartViewModel.CartListState.Result) {
@@ -140,25 +153,16 @@ class CartFragment : Fragment(R.layout.fragment_cart), ObserverListEmpty {
         }
     }
 
-    private fun setupEvents() = cartViewModel.apply {
-        getAllItemsFromDatabase()
+    private fun setupEvents() {
+        cartViewModel.getAllItemsFromDatabase()
+        preferencesManager = PreferencesManager.instance
+        intentSharedCart = Intent(Intent.ACTION_SEND)
     }
 
     private fun showDialogFeatureClearCart() {
         if (preferencesManager?.getViewFeatureClearCart() == false && preferencesManager?.getViewOnboardingWelcome() == true) {
             val dialog = FeatureClearCartDialog()
             dialog.show(childFragmentManager, Constants.DIALOG_FEATURE_CLEAR_CART)
-        }
-    }
-
-    private fun showList(isShow: Boolean) = binding.apply {
-        pbLoadingList.goneWithoutAnimation()
-        if (isShow) {
-            rcyCartItem.visibilityWithoutAnimation()
-            containerListEmpty.goneWithoutAnimation()
-        } else {
-            rcyCartItem.goneWithoutAnimation()
-            containerListEmpty.visibilityWithoutAnimation()
         }
     }
 
@@ -169,9 +173,15 @@ class CartFragment : Fragment(R.layout.fragment_cart), ObserverListEmpty {
             }
 
             if (isEmpty) {
-                showList(false)
+                binding.cartViewFlipper.displayedChild = VIEW_FLIPPER_EMPTY
                 setCartValue()
             }
         }
+    }
+
+    private companion object {
+        const val VIEW_FLIPPER_EMPTY = 0
+        const val VIEW_FLIPPER_LIST = 1
+        const val VIEW_FLIPPER_LOADING = 2
     }
 }
